refactor(ai): migrate list_chat_groups to tools registry
